// practices/button-generator/src/enums/button.ts
export enum SIZE {
  SMALL = '1',
  MEDIUM = '2',
  LARGE = '3',
}

export enum BORDER_SIZE {
  SMALL = '1',
  MEDIUM = '2',
  LARGE = '3',
  EXTRA_LARGE = '4',
}

export enum ROUNDED_SIZE {
  SMALL = '1',
  MEDIUM = '2',
  LARGE = '3',
  FULL = '4',
}

export enum VARIANT {
  PRIMARY = 'primary',
  SECONDARY = 'secondary',
  SUCCESS = 'success',
  DANGER = 'danger',
}